move permission flags to observer module
