--- proto/crosschain/permission_flags.proto
@@ -1,8 +1,0 @@
-syntax = "proto3";
-package zetachain.zetacore.crosschain;
-
-option go_package = "github.com/zeta-chain/zetacore/x/crosschain/types";
-
-message PermissionFlags {
-  bool isInboundEnabled = 1;
-}
