+++ proto/observer/genesis.proto
@@ -6,12 +6,14 @@ import "observer/observer.proto";
 import "observer/ballot.proto";
 import "observer/params.proto";
 import "observer/node_account.proto";
+import "observer/permission_flags.proto";
 
 option go_package = "github.com/zeta-chain/zetacore/x/observer/types";
 
 message GenesisState {
   repeated Ballot ballots =1;
   repeated ObserverMapper observers = 2;
-  repeated NodeAccount nodeAccountList = 8;
-  Params params = 3;
+  repeated NodeAccount nodeAccountList = 3;
+  PermissionFlags permissionFlags = 4;
+  Params params = 5;
 }
