+++ proto/observer/permission_flags.proto
@@ -1,0 +1,8 @@
+syntax = "proto3";
+package zetachain.zetacore.observer;
+
+option go_package = "github.com/zeta-chain/zetacore/x/observer/types";
+
+message PermissionFlags {
+  bool isInboundEnabled = 1;
+}
