+++ x/crosschain/keeper/in_tx_hash_to_cctx_test.go
@@ -1,6 +1,7 @@
 package keeper_test
 
 import (
+	"github.com/zeta-chain/zetacore/x/crosschain/keeper"
 	"strconv"
 	"testing"
 
@@ -8,7 +9,6 @@ import (
 	"github.com/stretchr/testify/require"
 	keepertest "github.com/zeta-chain/zetacore/testutil/keeper"
 	"github.com/zeta-chain/zetacore/testutil/nullify"
-	"github.com/zeta-chain/zetacore/x/crosschain/keeper"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
 )
