+++ x/crosschain/keeper/keeper_cross_chain_tx_vote_inbound_tx.go
@@ -31,7 +31,7 @@ import (
 func (k msgServer) VoteOnObservedInboundTx(goCtx context.Context, msg *types.MsgVoteOnObservedInboundTx) (*types.MsgVoteOnObservedInboundTxResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 	observationType := observerTypes.ObservationType_InBoundTx
-	if !k.IsInboundAllowed(ctx) {
+	if !k.zetaObserverKeeper.IsInboundAllowed(ctx) {
 		return nil, types.ErrNotEnoughPermissions
 	}
 	// GetChainFromChainID makes sure we are getting only supported chains , if a chain support has been turned on using gov proposal, this function returns nil
