+++ x/crosschain/types/genesis.go
@@ -16,7 +16,6 @@ func DefaultGenesis() *GenesisState {
 		// this line is used by starport scaffolding # ibc/genesistype/default
 		OutTxTrackerList:   []OutTxTracker{},
 		InTxHashToCctxList: []InTxHashToCctx{},
-		PermissionFlags:    nil,
 		// this line is used by starport scaffolding # genesis/types/default
 		Keygen:          nil,
 		GasPriceList:    []*GasPrice{},
