+++ x/observer/keeper/msg_server_permission_flags.go
@@ -7,7 +7,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/zeta-chain/zetacore/x/crosschain/types"
+	"github.com/zeta-chain/zetacore/x/observer/types"
 )
 
 // Updates permissions. Currently, this is only used to enable/disable the
@@ -16,7 +16,7 @@ import (
 // Only the admin policy account is authorized to broadcast this message.
 func (k msgServer) UpdatePermissionFlags(goCtx context.Context, msg *types.MsgUpdatePermissionFlags) (*types.MsgUpdatePermissionFlagsResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
-	if msg.Creator != k.zetaObserverKeeper.GetParams(ctx).GetAdminPolicyAccount(zetaObserverTypes.Policy_Type_stop_inbound_cctx) {
+	if msg.Creator != k.GetParams(ctx).GetAdminPolicyAccount(zetaObserverTypes.Policy_Type_stop_inbound_cctx) {
 		return &types.MsgUpdatePermissionFlagsResponse{}, zetaObserverTypes.ErrNotAuthorizedPolicy
 	}
 	// Check if the value exists
