+++ x/observer/types/genesis.go
@@ -14,6 +14,7 @@ func DefaultGenesis() *GenesisState {
 		Ballots:         nil,
 		Observers:       nil,
 		NodeAccountList: []*NodeAccount{},
+		PermissionFlags: &PermissionFlags{IsInboundEnabled: true},
 	}
 }
